Add files via upload
in this video, firstly i have created IAM(Identity and Access management) user group and given IAMReadonly acceess .        2. I have created user with IAMReadonlyaccess then attached user to the group        3. i have attached policies directly and via group to the user.                4. Created a Role (TO ATTACH THE ROLE TO THE EC2 INSTANCE) with IAMReadonlyrole and created an eC2 instance